add quickaction buttons, add clear sql fn
